Add madge to identify circular dependencies; fix those identified
