+++ packages/discovery-react-components/.madgerc
@@ -1,0 +1,4 @@
+{
+    "fileExtensions": ["js","ts","tsx"],
+    "tsConfig": "tsconfig.json"
+}
