+++ packages/discovery-react-components/package.json
@@ -11,6 +11,7 @@
     "test": "cross-env CI=1 react-scripts test",
     "test:watch": "react-scripts test --watch",
     "test:debug": "react-scripts --inspect-brk test --runInBand --no-cache",
+    "circular-dependencies": "madge --circular src/*",
     "code-coverage": "jest --collectCoverage",
     "build": "rollup -c",
     "start": "rollup -c -w",
