+++ packages/discovery-react-components/src/components/CIDocument/components/MetadataPane/components/MetadataRow.tsx
@@ -1,11 +1,8 @@
 import React, { FC } from 'react';
 import cx from 'classnames';
 import { Link, Tooltip } from 'carbon-components-react';
-import {
-  Metadata,
-  MetadataData,
-  OnActiveMetadataChangeFn
-} from '@CIDocument/components/MetadataPane/types';
+import { Metadata, MetadataData } from '@CIDocument/types';
+import { OnActiveMetadataChangeFn } from '@CIDocument/components/MetadataPane/types';
 import { getId } from '@rootUtils/document/idUtils';
 import { ReactComponent as HighConfidence } from '@CIDocument/components/MetadataPane/components/icons/Confidence_high.svg';
 import { ReactComponent as MediumConfidence } from '@CIDocument/components/MetadataPane/components/icons/Confidence_medium.svg';
